Title feat: add todo creation functionality and update views
Overview
Added a new route and controller logic for creating and storing todos.
Updated the navigation to include a link to the todo creation page.
Enhanced the Todo model to include user_id in the fillable attributes.
Modified the todos listing to show the latest todos first.
Updated the todo creation view and removed debug code.
Minor formatting adjustments in the todo show view.